I don't remember

+++ OpenCLPrograms/tools.cl
@@ -134,9 +134,9 @@ float3 generate_hemisphere_with_solid_angle(t_random *random, float angle_max)
     float r1 = randf(random);
     float r2 = randf(random);
     return (float3){
-        cosf(2.0f * M_PI_F * r1) * sqrtf(1.0f - powf(1.0f - r2 * (1.0f - cosf(angle_max)), 2.0f)),
-        sinf(2.0f * M_PI_F * r1) * sqrtf(1.0f - powf(1.0f - r2 * (1.0f - cosf(angle_max)), 2.0f)),
-        1.0f - r2 * (1.0f - cosf(angle_max))
+        cos(2.0f * M_PI_F * r1) * sqrt(1.0f - pow(1.0f - r2 * (1.0f - cos(angle_max)), 2.0f)),
+        sin(2.0f * M_PI_F * r1) * sqrt(1.0f - pow(1.0f - r2 * (1.0f - cos(angle_max)), 2.0f)),
+        1.0f - r2 * (1.0f - cos(angle_max))
     };
 }
 
@@ -195,9 +195,9 @@ float3 random_point_on_sphere(__global t_sphere *sphere, t_random *random)
     float phi = 2*M_PI_F*r2;
 
     return (float3){
-        sinf(theta) * cosf(phi),
-        cosf(theta),
-        sinf(theta) * sin(phi)
+        sin(theta) * cos(phi),
+        cos(theta),
+        sin(theta) * sin(phi)
     } + sphere->position;
 }
